Removed duplicated files
