+++ app/src/androidTest/java/com/github/ravifrancesco/softwaredevproject/TopographyAsyncTest.java
@@ -11,14 +11,15 @@ import org.junit.Test;
 import org.osmdroid.util.GeoPoint;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 
 public class TopographyAsyncTest {
 
     private static Pair<int[][], Double> topographyPair;
     private static UserPoint userPoint;
-    private ElevationMapAsync elevationMapAsync;
 
     /**
      * Download the topography map
@@ -53,14 +54,14 @@ public class TopographyAsyncTest {
     }
 
     /**
-     * Tests the ElevationMapAsyncClass when the topographyMap is null
+     * Tests the elevationMapClass when the topographyMap is null
      */
     @Test
     public void topographyNull(){
         Pair<int[][], Double> topo = new Pair<>(null, 0.0);
-        elevationMapAsync = new ElevationMapAsync(topo, userPoint);
-        Assert.assertNull(elevationMapAsync.getIndexesFromCoordinates(userPoint.getLatitude(), userPoint.getLongitude()));
-        Assert.assertEquals(0, elevationMapAsync.getAltitudeAtLocation(userPoint.getLatitude(), userPoint.getLongitude()));
+        ElevationMap elevationMapNull = new ElevationMap(topo, userPoint);
+        Assert.assertNull(elevationMapNull.getIndexesFromCoordinates(userPoint.getLatitude(), userPoint.getLongitude()));
+        Assert.assertEquals(0, elevationMapNull.getAltitudeAtLocation(userPoint.getLatitude(), userPoint.getLongitude()));
     }
 
     /**
@@ -70,12 +71,68 @@ public class TopographyAsyncTest {
     public void topographyNotNull(){
         Assert.assertNotNull(topographyPair);
         Assert.assertNotNull(topographyPair.first);
-        elevationMapAsync = new ElevationMapAsync(topographyPair, userPoint);
-        Assert.assertArrayEquals(topographyPair.first, elevationMapAsync.getTopographyMap());
+        ElevationMap elevationMap = new ElevationMap(topographyPair, userPoint);
+        Assert.assertArrayEquals(topographyPair.first, elevationMap.getTopographyMap());
     }
 
     /**
-     * Same test as @see ElevationMapTest.getAltitudeTest.java but with ElevationMapAsync
+     * Check that indexes for accessing the matrix are computed correctly
+     */
+    @Test
+    public void getIndexesTest() {
+
+        Context mContext = ApplicationProvider.getApplicationContext();
+
+        UserPoint userPoint = new UserPoint(mContext);
+        userPoint.setLocation(GPSTracker.DEFAULT_LAT, GPSTracker.DEFAULT_LON,GPSTracker.DEFAULT_ALT, GPSTracker.DEFAULT_ACC);
+
+        ElevationMap elevationMap = new ElevationMap(topographyPair, userPoint);
+
+        // check the indexes around the Everest Peak
+        Assert.assertEquals(new Pair<>(215, 244), elevationMap.getIndexesFromCoordinates(GPSTracker.DEFAULT_LAT, GPSTracker.DEFAULT_LON));
+
+        // set location near the Mont Blanc
+        userPoint.setLocation(45.802537, 6.850328, 0, 0);
+        elevationMap.updateElevationMatrix();
+
+        // check the indexes around the Mont Blanc Peak
+        Assert.assertEquals(new Pair<>(178, 326), elevationMap.getIndexesFromCoordinates(45.8326, 6.8652));
+    }
+
+    /**
+     * Test that the map cell size is computed correctly
+     */
+    @Test
+    public void getMapCellSizeTest() {
+
+        Context mContext = ApplicationProvider.getApplicationContext();
+
+        UserPoint userPoint = new UserPoint(mContext);
+        userPoint.setLocation(GPSTracker.DEFAULT_LAT, GPSTracker.DEFAULT_LON,GPSTracker.DEFAULT_ALT, GPSTracker.DEFAULT_ACC);
+
+        ElevationMap elevationMap = new ElevationMap(topographyPair, userPoint);
+
+        Assert.assertEquals(0.000833333333, elevationMap.getMapCellSize(), 0.00000001);
+    }
+
+    /**
+     * Test that the bounding box is returned correctly
+     */
+    @Test
+    public void getBoundingBoxWestLongTest() {
+
+        Context mContext = ApplicationProvider.getApplicationContext();
+
+        UserPoint userPoint = new UserPoint(mContext);
+        userPoint.setLocation(GPSTracker.DEFAULT_LAT, GPSTracker.DEFAULT_LON,GPSTracker.DEFAULT_ALT, GPSTracker.DEFAULT_ACC);
+
+        ElevationMap elevationMap = new ElevationMap(topographyPair, userPoint);
+
+        Assert.assertEquals(userPoint.computeBoundingBox(ElevationMap.BOUNDING_BOX_RANGE).getLonWest(), elevationMap.getBoundingBoxWestLong(), 0.00000001);
+    }
+
+    /**
+     * Checks if the altitude is calculated correctly and that the elevation map gets udated correctly
      */
     @Test
     public void getAltitudeTest() throws InterruptedException {
@@ -85,7 +142,7 @@ public class TopographyAsyncTest {
         UserPoint userPoint = new UserPoint(mContext);
         userPoint.setLocation(GPSTracker.DEFAULT_LAT, GPSTracker.DEFAULT_LON,GPSTracker.DEFAULT_ALT, GPSTracker.DEFAULT_ACC);
 
-        ElevationMapAsync elevationMap = new ElevationMapAsync(topographyPair, userPoint);
+        ElevationMap elevationMap = new ElevationMap(topographyPair, userPoint);
 
         // check the altitude around the Everest Peak using coordinates
         Assert.assertEquals(8849, elevationMap.getAltitudeAtLocation(GPSTracker.DEFAULT_LAT, GPSTracker.DEFAULT_LON), 200);
@@ -120,13 +177,12 @@ public class TopographyAsyncTest {
     }
 
     /**
-     * Test LineOfSightAsync.getVisiblePoints to see if the POIPoints are correctly filtered
-     * @see LineOfSightTest
-     *
+     * This tests checks if the POIPoints are filtered correctly by the
+     * getVisiblePoints method.
      */
     @Test
     public void getVisiblePoints(){
-        LineOfSightAsync lineOfSight = new LineOfSightAsync(topographyPair, userPoint);
+        LineOfSight lineOfSight = new LineOfSight(topographyPair, userPoint);
 
         // PoiPoints to check
         List<POIPoint> pointsToCheck = new ArrayList<>();
@@ -154,5 +210,49 @@ public class TopographyAsyncTest {
 
         // Check if the points are filtered correctly
         Assert.assertEquals(new HashSet<>(lineOfSight.getVisiblePoints(pointsToCheck)), new HashSet<>(visiblePoints));
+    }
+
+    /**
+     * This tests checks if the POIPoints are labeled correctly by the
+     * getVisiblePointsLabeled method.
+     */
+    @Test
+    public void getVisiblePointsLabeledTest() {
+
+        Context mContext = ApplicationProvider.getApplicationContext();
+
+        // setting location near everest peak
+        UserPoint userPoint = new UserPoint(mContext);
+        userPoint.setLocation(28.000490, 86.921267,7500, 0);
+
+        LineOfSight lineOfSight = new LineOfSight(topographyPair, userPoint);
+
+        // PoiPoints to check
+        List<POIPoint> pointsToCheck = new ArrayList<>();
+        POIPoint point1 = new POIPoint(new GeoPoint(28.011581, 86.907036, 6200));
+        POIPoint point2 = new POIPoint(new GeoPoint(28.017394, 86.922196, 7000));
+        point1.setName("point1");
+        point2.setName("point2");
+        // points that should not be visible
+        POIPoint point3 = new POIPoint(new GeoPoint(27.951538, 86.928781, 6400));
+        POIPoint point4 = new POIPoint(new GeoPoint(27.987947, 86.933671, 8000));
+        point3.setName("point3");
+        point4.setName("point4");
+
+        // Add the points to the points to check
+        pointsToCheck.add(point1);
+        pointsToCheck.add(point2);
+        pointsToCheck.add(point3);
+        pointsToCheck.add(point4);
+
+        // Create the test map
+        Map<POIPoint, Boolean> labeledPOIPoints = new HashMap<>();
+        labeledPOIPoints.put(point1, true);
+        labeledPOIPoints.put(point2, true);
+        labeledPOIPoints.put(point3, false);
+        labeledPOIPoints.put(point4, false);
+
+        // Check if the points are labeled correctly
+        Assert.assertEquals(labeledPOIPoints, lineOfSight.getVisiblePointsLabeled(pointsToCheck));
     }
 }
